Added Login Functionality
Added login functionality by calling the api endpoint and checking if
user exists for a `username` and `password`.
The credentials (required for api call) are stored in a `.env` file.
Logged in user's details are now available for the parent as well as
sibling components.
Future: Try to maintain login session, so that on-reload the user does
not logs out